# public/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Piyachok API",
    "description": "Документация backend-платформы проекта Piyachok."
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Введіть токен у форматі: Bearer <AccessToken>"
    }
  },
  "paths": {
...
    "/venues/{id}": {
      "get": {
        "tags": ["Venues"],
        "summary": "Отримання детальної інформації про заклад",
        "description": "Повертає повну інформацію про заклад за його ID, включаючи теги, відсортовані за датою новини та відгуки з іменами авторів. Якщо запит робить авторизований користувач, додатково повертається прапорець `isFavorite`. Доступно публічно.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID закладу"
          }
        ],
        "responses": {
          "200": {
            "description": "Дані закладу успішно отримано",
            "schema": {
              "type": "object",
              "properties": {
                "venue": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "758276ea-f44c-4cd9-9545-85778bd6b847" },
                    "name": { "type": "string", "example": "П'яний Лис" },
                    "type": { "type": "string", "example": "PUB" },
                    "address": { "type": "string", "example": "вул. Садова, 12" },
                    "views": { "type": "integer", "example": 142, "description": "Загальна кількість переглядів" },
                    "isFavorite": { "type": "boolean", "example": false, "description": "Чи додано заклад до обраного поточним користувачем" },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "tag-1" },
                          "name": { "type": "string", "example": "Крафт" }
                        }
                      }
                    },
                    "news": { "type": "array", "items": { "type": "object" } },
                    "reviews": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заклад не знайдено",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      },
      "patch": {
        "tags": ["Venues"],
        "summary": "Редагування даних закладу",
        "description": "Дозволяє оновити будь-які поля закладу. Доступно тільки власнику закладу (менеджеру) або супер-адміністратору. Валідується за допомогою схеми.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID закладу"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "Нова назва закладу" },
                "address": { "type": "string", "example": "Нова адреса, 45" },
                "phone": { "type": "string", "x-nullable": true, "example": "+380481234567" },
                "averageCheck": { "type": "integer", "example": 350 },
                "hasWifi": { "type": "boolean", "example": true },
                "tags": {
                  "type": "array",
                  "items": { "type": "string" },
                  "example": ["Крафт", "Укриття", "Тераса"]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заклад успішно відредаговано",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Заклад успішно відредаговано!" },
                "venue": { "type": "object" }
              }
            }
          },
          "400": {
            "description": "Помилка валідації вхідних даних",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "401": {
            "description": "Неавторизований доступ",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "403": {
            "description": "Недостатньо прав для редагування (користувач не є власником)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Заклад не знайдено",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      },
      "delete": {
        "tags": ["Venues"],
        "summary": "Видалення закладу",
        "description": "Повністю видаляє заклад та пов'язані з ним каскадні дані з бази даних. Доступно тільки власнику закладу або супер-адміністратору.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID закладу, який потрібно видалити"
          }
        ],
        "responses": {
          "200": {
            "description": "Заклад успішно видалено",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Заклад та всі пов’язані з ним дані успішно видалено!" }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "403": {
            "description": "Недостатньо прав для видалення",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Заклад не знайдено",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": { "type": "string", "example": "Внутрішня помилка сервера / Доступ заборонено" }
      }
    },
...
  }
}